Fix CO2 display range comparison

+++ views/General_view.c
@@ -177,7 +177,7 @@ static void _draw_co2(Canvas* canvas, Sensor* sensor, Color color) {
     //Drawing icon
     canvas_draw_icon(canvas, x + 3, y + 3, &I_co2_11x14);
 
-    int16_t concentration_int = sensor->co2;
+    int32_t concentration_int = (int32_t)sensor->co2;
 //    int8_t concentration_dec = (int16_t)(sensor->co2 * 10) % 10;
 
     //Whole part
